pip-requirements.txt
django==3.0.1
pytz==2019.3
djangorestframework==3.11.0
psycopg2==2.8.4
django-rest-auth==0.9.5
six==1.13.0
django-allauth==0.41.0
python3-openid==3.1.0
oauthlib==3.1.0
requests==2.22.0
requests-oauthlib==1.3.0
defusedxml==0.6.0
idna==2.8
urllib3==1.25.7
chardet==3.0.4
certifi==2019.11.28
aylien-apiClient==0.7.0
httplib2==0.15.0
Cython==0.29.14